+++ src/main/scala/org/bitcoins/spvnode/messages/data/MerkleBlockMessage.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.spvnode.messages.data
 
 import org.bitcoins.core.crypto.DoubleSha256Digest
+import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.CompactSizeUInt
 import org.bitcoins.core.protocol.blockchain.BlockHeader
 import org.bitcoins.core.util.Factory
@@ -13,13 +14,13 @@ import org.bitcoins.spvnode.serializers.messages.data.RawMerkleBlockMessageSeria
   */
 object MerkleBlockMessage extends Factory[MerkleBlockMessage] {
 
-  private case class MerkleBlockMessageImpl(blockHeader : BlockHeader, transactionCount : Long,
+  private case class MerkleBlockMessageImpl(blockHeader : BlockHeader, transactionCount : UInt32,
                                             hashCount : CompactSizeUInt, hashes : Seq[DoubleSha256Digest],
                                             flagCount : CompactSizeUInt, flags : Seq[Byte]) extends MerkleBlockMessage
 
   def fromBytes(bytes : Seq[Byte]) : MerkleBlockMessage = RawMerkleBlockMessageSerializer.read(bytes)
 
-  def apply(blockHeader: BlockHeader, transactionCount : Long, hashCount : CompactSizeUInt, hashes : Seq[DoubleSha256Digest],
+  def apply(blockHeader: BlockHeader, transactionCount : UInt32, hashCount : CompactSizeUInt, hashes : Seq[DoubleSha256Digest],
             flagCount : CompactSizeUInt, flags : Seq[Byte]) : MerkleBlockMessage = {
     MerkleBlockMessageImpl(blockHeader, transactionCount,hashCount,hashes,flagCount,flags)
   }
